Updated import paths for moved packages
